requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
tomli==2.0.1
tomli-w==1.0.0
